GTM19-542: Python 3.7 support (release branch only commit)
Fix to get gdb to build with a cross-compiled Python (for mingw).
Parts of this may be suitable for upstreaming.  The original configure
check for python relies on a working python executable which is not
valid for cross builds.  This change adds the capability to use the
python-config script which works even for mingw hosted toolchains
cross-built on linux.

gdb/
	* configure.ac: Add support for python 3.7 and the capability
	to derive compiler/linker flags from python-config utility.
	* configure: Regenerate using autoconf 2.69
	* python/python.c: Fix #ifdef conditional for MINGW builds.