CR MOLY00715449: Enable relaxation of 32-bit branches at the range boundary
Branches that lie at the boundary of 2 and 4 byte ranges can theoretically
cause the relaxation loop to oscillate between the two options.  It is not
possible to predict the exact oscillation condition so we approximate it by
examining branches that get expanded from 2 to 4 bytes at the range boundary
in any iteration.  Forcing these expansions to be permanent avoids
oscillation while still allowing 2-byte branches to utilize their full
valid range in *most* cases. Some branches will still get expanded to 32-bit
where not necessary but solving those is NP.

The earlier solution was to assume that all 16-bit branches at their boundary
are to be expanded.  We forego the final 2 bytes of the 16-bit range and
generate 32-bit branches even when not strictly required.

gas/
	* tc-nanomips.c (RELAX_MD_TOOFAR_FINAL): New.
	(RELAX_MD_MARK_TOOFAR_FINAL): New.
	(relaxed_16bit_branch_length): Mark frag as final when expanding to
	4-byte branch at the boundary condition. Skip final frags on subsequent
	iterations.
	* testsuite/gas/nanomips/branch-relax.s: Update tests for boundary
	conditions and add more cases.
	* testsuite/gas/nanomips/branch-relax.d: Update reference output.
	* testsuite/gas/nanomips/branch-relax32.d: Update reference output.
	* testsuite/gas/nanomips/branch-relax.l: Update reference output.
	* testsuite/gas/nanomips/branch-relax2.s: New source file.
	* testsuite/gas/nanomips/branch-relax2.s: New test.
	* testsuite/gas/nanomips/nanomips.exp: Run the new test.

Change-Id: I11a171d5d3561b96d4aaf416631851d8b9c7298e