BZ#13: Reduce differences between builds toggling minimize-relocs
GOLD reaches a different outcome from GAS on certain branch relations,
specifically where the branch is exactly on the boundary of possible 2
and 4 byte ranges.  The assembler is conservative in picking the 4-byte
instruction where the linker is able to pick 2-byte will information
available from the link time address layout.  As a results build with
and without minimize-relocs give difference final linked outputs.

This assembler change tries to reduce such differences by punting in
favor of the linker and skipping them for the minimize-relocs pass.
Assembler retains relocations on such branch instructions, so that the
linker can minimize them where possible.

gas/
        * config/tc-nanomips.c (mark_if_variable_branch): Skip branch
        instructions at the boundary of 2 and 4-byte branch ranges from
        the minimize-relocs pass.
	* testsuite/gas/nanomips/minimize-relocs8.s: New source.
	* testsuite/gas/nanomips/minimize-relocs8.d: New test.
	* testsuite/gas/nanomips/nanomips.exp: Run the new test.

Change-Id: I5d09712dd4e169d01a0ef733e8ec1eac5616b2bf